Move eth directory to market-closure/helpers
