+++ Makefile
@@ -21,11 +21,9 @@ clean-market-closure:
 build-market-closure:
 	cp $(adapter)/adapter.js market-closure/$(check)/priceAdapter.js
 	cp market-closure/adapter.js market-closure/$(check)
-	cp eth/readReferenceContract.js market-closure/$(check)
 	yarn ncc build market-closure/$(check) -o market-closure/$(check)/dist
 	rm market-closure/$(check)/priceAdapter.js
 	rm market-closure/$(check)/adapter.js
-	rm market-closure/$(check)/readReferenceContract.js
 
 docker-market-closure:
 	docker build --no-cache --build-arg adapter=$(adapter) --build-arg check=$(check) -f Dockerfile-MarketClosure . -t $(adapter)-$(check)-adapter
