+++ market-closure/helpers/eth/readReferenceContract.js
